Photo controller and service added

+++ backend/src/main/java/com/alper/shotify/backend/repository/IPhotoRepository.java
@@ -3,5 +3,5 @@ package com.alper.shotify.backend.repository;
 import com.alper.shotify.backend.entity.PhotoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPhotoRepository extends JpaRepository<PhotoEntity, Long> {
+public interface IPhotoRepository extends JpaRepository<PhotoEntity, Integer> {
 }
